reports-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
clickhouse-driver==0.2.6
python-jose[cryptography]==3.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
boto3==1.34.0
python-multipart==0.0.6
